+++ src/backend/tests/services/test_compass.py
@@ -1,0 +1,10 @@
+from backend.services.compass import Compass
+
+
+def test_compass() -> None:
+    compass = Compass()
+    result = compass.invoke(
+        action=Compass.ValidActions.CREATE_INDEX, parameters={"index": "foobar"}
+    )
+    assert result.result is None
+    assert result.result is None
